Added MacOS Support

+++ include/debug.h
@@ -6,7 +6,7 @@
 namespace neutron {
 
 void disassembleChunk(const Chunk* chunk, const char* name);
-int disassembleInstruction(const Chunk* chunk, int offset);
+size_t disassembleInstruction(const Chunk* chunk, size_t offset);
 
 } // namespace neutron
